NP-46465 Change from hook to wrapper to ensure default values are added everytime one visits ticket page (#5393)
* NP-46465 Move defaultParams hook from TasksPAge into TicketList to ensure it runs everytime onw visits ticket page

* NP-46465 MVP default ticket search - remember params

* Use wrapper instead of hook to add default values

* Import order

* Import order (again)

* Set LocationStatus type on useLocation instead of inline
